Add a used hover hook for the sbout psgr
